+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 ## Changes compared to v2.0.0.5601_dev
 
 - I'm doing my best to optimise EmuGUI by adapting the New VM dialog.
+- For that, the plug-in reader has been prepared.
 
 ## Known issues